E-BPR/requirements.txt
absl-py==2.1.0
addict==2.4.0
appdirs==1.4.4
attrs==24.2.0
cachetools==5.5.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.4.1
cityscapesScripts==2.2.4
click==8.1.8
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.2.7
cycler==0.11.0
exceptiongroup==1.2.2
fire==0.7.0
flake8==5.0.4
fonttools==4.38.0
future==1.0.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
humanfriendly==10.0
idna==3.10
importlib-metadata==6.7.0
iniconfig==2.0.0
interrogate==1.5.0
isort==4.3.21
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mccabe==0.7.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mmcv-full==1.2.0
-e git+ssh://git@github.com/JadeHannah/RefineMask.git@f9f68c6a9188c827d74c41750f85c487cb6eb1ee#egg=mmsegmentation
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.0
Pillow==9.4.0
platformdirs==4.0.0
pluggy==1.2.0
protobuf==3.20.3
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==7.4.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.7.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.67.1
typing==3.7.4.3
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
xdoctest==1.1.6
yapf==0.40.0
zipp==3.15.0
